Updated HTTPEndpointHandler Response Body to Int64 from Int. Minor optimization in SQLiteResultsRow. Added sum(...) to SQLiteTableColumn. Reorganized method order in SQLiteTable. Added migrate(...) and sum(...) to SQLiteTable.

+++ Source/Networking/HTTPEndpointHandler.swift
@@ -53,7 +53,7 @@ public enum HTTPEndpointResponseBody {
 
 	// MARK: Values
 	case data(_ value :Data)
-	case integer(_ value :Int)
+	case integer(_ value :Int64)
 	case json(_ value :Any)
 	case string(_ value :String)
